Python/Flask/Flask_Restful_API/02_BasicCrudImplent_Addictional_Resource/requirements.txt
Flask==0.12
Flask-HTTPAuth==2.2.1
Flask-SQLAlchemy==2.1
Jinja2==2.9.6
MarkupSafe==1.0
Pygments==2.2.0
SQLAlchemy==1.0
Werkzeug==0.9.6
httpie==0.9.9
itsdangerous==0.24
python-dateutil==2.6.1
requests==2.18.4
six==1.10.0
coverage==3.7.1